page numbers helper Extracts page naming logic from renderer code
